+++ torch/_dynamo/variables/__init__.py
@@ -110,28 +110,31 @@ from .lists import (
     TupleIteratorVariable,
     TupleVariable,
 )
+from .descriptors import (
+    BuiltinMethodVariable,
+    ClassMethodDescriptorVariable,
+    ClassMethodVariable,
+    GetSetDescriptorVariable,
+    MemberDescriptorVariable,
+    MethodDescriptorVariable,
+    MethodWrapperVariable,
+    PropertyVariable,
+    StaticMethodVariable,
+    WrapperDescriptorVariable,
+)
 from .misc import (
     AutogradFunctionContextVariable,
     AutogradFunctionVariable,
-    BuiltinMethodVariable,
     CellVariable,
-    ClassMethodDescriptorVariable,
-    ClassMethodVariable,
     ConstantMethodWrapperVariable,
     DeletedVariable,
     ExceptionVariable,
     GetAttrVariable,
-    GetSetDescriptorVariable,
     LambdaVariable,
-    MemberDescriptorVariable,
-    MethodDescriptorVariable,
-    MethodWrapperVariable,
     NewGlobalVariable,
     NumpyVariable,
     ObjectVariable,
-    PropertyVariable,
     PythonModuleVariable,
-    StaticMethodVariable,
     RandomClassVariable,
     RandomVariable,
     StringFormatVariable,
@@ -141,7 +144,6 @@ from .misc import (
     TypingVariable,
     UnknownVariable,
     WeakRefVariable,
-    WrapperDescriptorVariable,
 )
 from .nn_module import (
     FSDPManagedNNModuleVariable,
